# specs/api.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Api pour l'application Goofood
paths:
...
  /api/utilisateur:
    get:
      security:
        - bearerAuth: []
      summary: une route qui permet de chercher un utilisateur
      tags:
        - Utilisateur
      responses:
        "200":
          description: les données d'un utilisateur sont retournées
          content:
            application/json:
              schema:
                type: object
                required:
                  - idUtilisateur
                  - email
                  - nom
                  - photoDeProfile
                  - noTelephone
                  - adresse
                properties:
                  idUtilisateur:
                    type: number
                  email:
                    type: string
                  nom:
                    type: string
                  photoDeProfile:
                    type: string
                  noTelephone:
                    type: string
                  adresse:
                    type: object
                    required:
                      - noRue
                      - nomRue
                      - nomVille
                      - province
                      - codePostal
                    properties:
                      noRue:
                        type: string
                      nomRue:
                        type: string
                      nomVille:
                        type: string
                      province:
                        type: string
                      codePostal:
                        type: string
        "404":
          description: L'utilisateur n'existe pas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
    put:
      security:
        - bearerAuth: []
      summary: route pour modifier les données d'un utilisateur
      tags:
        - Utilisateur
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                motDePasse:
                  type: string
                nom:
                  type: string
                photoDeProfile:
                  type: string
                noTelephone:
                  type: string
      responses:
        "200":
          description: les données de l'utilisateur ont été modifiées
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Modification"
    delete:
      security:
        - bearerAuth: []
      summary: Route pour suprimer un utilisateur
      tags:
        - Utilisateur
      responses:
        "200":
          description: L'user est supprimé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Delete"
        "400":
          description: Élément non trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/updateProfiles:
    put:
      security:
        - bearerAuth: []
      summary: Route pour changer une photo de profiles d'un utilisateur
      tags:
        - Utilisateur
      requestBody:
        content:
          'image/jpeg':
            schema:
              type: string
              format: binary
          'image/png':
            schema:
              type: string
              format: binary  
      responses:
        "200":
          description: Photo de profile changé
          content:
            application/json:
              schema:
                type: object
                properties:
                  urlProfile:
                    type: string
        "400":
          description:  fichier non traité
          content:
            application/json:
              schema:
                type: object
                properties:
                  erreur:
                    type: string
  /api/restaurant/updatelogoRestaurant:
    put:
      tags:
        - Restaurant
      parameters:
        - in: query
          name: idRestaurant
          schema:
            type: number
          required: true
          description: Id du restaurant
      requestBody:
        content:
            # a binary file of type png or jpeg
          'image/jpeg':
            schema:
              type: string
              format: binary
          'image/png':
            schema:
              type: string
              format: binary  
      responses:
        "200":
          description: logo changé
          content:
            application/json:
              schema:
                type: object
                properties:
                  urllogoImage:
                    type: string
        "404":
          description: fichier non traité
          content:
            application/json:
              schema:
                type: object
                properties:
                  erreur:
                    type: string
...
  /api/restaurant:
    get:
      summary: Route pour obtenir les données d'un restaurants
      tags:
        - Restaurant
      parameters:
        - in: query
          name: idRestaurant
          schema:
            type: number
          required: true
          description: Id du restaurant
      responses:
        "200":
          description: les informations d'un restaurant sont retournées
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "404":
          description: les données ne sont pas retournées
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
  /api/utilisateur/adresse:
    put:
      security:
        - bearerAuth: []
      tags:
        - Utilisateur
      summary: Route pour modifier une adresse d'un utilisateur
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Adresse"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Adresse"
        "400":
          description: impossible de changer l'adresse
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/restaurant/adresse:
    put:
      tags:
        - Restaurant
      summary: Route pour modifier une adresse d'un Restaurant
      parameters:
        - in: query
          name: idRestaurant
          schema:
            type: number
          required: true
          description: Id du restaurant
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Adresse"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Adresse"
        "400":
          description: impossible de changer l'adresse
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Modification:
      type: object
      properties:
        modification:
          type: boolean
    Delete:
      type: object
      properties:
        Delete:
          type: boolean
    Adresse:
      type: object
      properties:
        noRue:
          type: string
        nomRue:
          type: string
        nomVille:
          type: string
        province:
          type: string
        codePostal:
          type: string
...
    Restaurant:
      type: object
      required:
        - idRestaurant
        - nom
        - logoImage
        - noTelephone
        - description
        - capaciterMaximal
        - menu
        - category
        - adresse
      properties:
        idRestaurant:
          type: number
        noTelephone: 
          type: string
        nom:
          type: string
        logoImage:
          type: string
        description:
          type: string
        capaciterMaximal:
          type: string
        menu:
          type: object
          required:
            - SectionAliments
          properties:
            sectionAliments:
              type: array
              items:
                type: object
                required:
                  - idSectionAliment
                  - nom
                  - aliments
                properties:
                  idSectionAliment:
                    type: number
                  nom:
                    type: string
                  aliments:
                    type: array
                    items:
                      type: object
                      required:
                        - idAliment
                        - nom
                        - prix
                        - quantiter
                        - description
                      properties:
                        idAliment:
                          type: number
                        nom:
                          type: string
                        prix:
                          type: number
                        quantiter:
                          type: number
                        description:
                          type: string
        category:
          type: object
          required:
            - idCategoriesRestaurant
            - nom
          properties:
            idCategoriesRestaurant:
              type: number
            nom:
              type: string
        adresse:
          type: object
          required:
            - noRue
            - nomRue
            - nomVille
            - province
            - codePostal
          properties:
            noRue:
              type: string
            nomRue:
              type: string
            nomVille:
              type: string
            province:
              type: string
            codePostal:
              type: string
...
    NotFound:
      type: object
      required:
        - statusError
      properties:
        statusError:
          type: string